[refac] refactor package structure for student record metrics
